define class and methods to use the Shunting Yard Algorithm for parsing instead of AST
It is very difficult to build an Abstract Syntax Tree directly from tokens because of their infix notation. The solution to that problem is convert it into a postfix notation (also known as reverse polish notation) and then build the AST.

I'm getting rid of AST entirely since the tokens in postfix notation would already be quite easier to evaluate. Using an AST would just add unnecessary complexity in code.